twitter-mood-detector/requirements.txt
tweepy>=4.14.0
pandas>=2.2.0
scikit-learn>=1.3.0
nltk>=3.8.0
transformers>=4.36.0
torch>=2.1.0
streamlit>=1.29.0
python-dotenv>=1.0.0
textblob>=0.17.0
vaderSentiment>=3.3.0
wordcloud>=1.9.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
numpy>=1.24.0
